tutorial command refactor to seperate files
