+++ packages/apis/src/CeramicApi.ts
@@ -4,6 +4,13 @@ import KeyDidResolver from 'key-did-resolver';
 import { DID } from 'dids';
 import { Ed25519Provider } from 'key-did-provider-ed25519';
 
+export type TutorialContent = {
+  name: string;
+  path: string;
+  digest: string;
+  arweaveHash?: string;
+}
+
 export type TutorialMetadata = {
   title: string;
   slug: string;
@@ -12,14 +19,7 @@ export type TutorialMetadata = {
   tags: string[];
   publishedUri?: string;
   createdAt: number;
-  content?: {
-    [filename: string]: {
-      name: string;
-      path: string;
-      digest: string;
-      arweaveHash?: string;
-    }
-  };
+  content?: { [filename: string]: TutorialContent };
 };
 
 export type ApiConfig = {
