+++ packages/cli/tsconfig.json
@@ -1,5 +1,8 @@
 {
   "extends": "@builderdao/tsconfig/react-library.json",
+  "compilerOptions": {
+    "baseUrl": "."
+  },
   "include": ["."],
   "exclude": ["dist", "build", "node_modules"]
 }
